// packages/solid-webgpu-gltf/src/generated/glTF.ts
export type GlTFChildOfRootProperty = GlTFProperty & {
  /**
   * The user-defined name of this object.
   */
  name?: string
  [k: string]: unknown
}
export type GlTFId = number
// ...
export type Animation = GlTFChildOfRootProperty & {
  /**
   * An array of animation channels. An animation channel combines an animation sampler with a target property being animated. Different channels of the same animation **MUST NOT** have the same targets.
   *
   * @minItems 1
   */
  channels: AnimationChannel[]
  /**
   * An array of animation samplers. An animation sampler combines timestamps with a sequence of output values and defines an interpolation algorithm.
   *
   * @minItems 1
   */
  samplers: AnimationSampler[]
  name?: unknown
  extensions?: unknown
  extras?: unknown
  [k: string]: unknown
}
/**
 * An animation channel combines an animation sampler with a target property being animated.
 */
export type AnimationChannel = GlTFProperty & {
  /**
   * The index of a sampler in this animation used to compute the value for the target.
   */
  sampler: GlTFId
  /**
   * The descriptor of the animated property.
   */
  target: AnimationChannelTarget
  extensions?: unknown
  extras?: unknown
  [k: string]: unknown
}
/**
 * The descriptor of the animated property.
 */
export type AnimationChannelTarget = GlTFProperty & {
  /**
   * The index of the node to animate. When undefined, the animated object **MAY** be defined by an extension.
   */
  node?: GlTFId
  /**
   * The name of the node's TRS property to animate, or the `"weights"` of the Morph Targets it instantiates. For the `"translation"` property, the values that are provided by the sampler are the translation along the X, Y, and Z axes. For the `"rotation"` property, the values are a quaternion in the order (x, y, z, w), where w is the scalar. For the `"scale"` property, the values are the scaling factors along the X, Y, and Z axes.
   */
  path: 'translation' | 'rotation' | 'scale' | 'weights' | string
  extensions?: unknown
  extras?: unknown
  [k: string]: unknown
}
/**
 * An animation sampler combines timestamps with a sequence of output values and defines an interpolation algorithm.
 */
export type AnimationSampler = GlTFProperty & {
  /**
   * The index of an accessor containing keyframe timestamps.
   */
  input: GlTFId
  /**
   * Interpolation algorithm.
   */
  interpolation?: 'LINEAR' | 'STEP' | 'CUBICSPLINE' | string
  /**
   * The index of an accessor, containing keyframe output values.
   */
  output: GlTFId
  extensions?: unknown
  extras?: unknown
  [k: string]: unknown
}
// ...
export interface GlTFProperty {
  extensions?: Extension
  extras?: Extras
  [k: string]: unknown
}